Move AndroidKeyEvents into helper class
