+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/handlers/SetDate.java
@@ -17,7 +17,10 @@
 package io.appium.espressoserver.lib.handlers;
 
 import androidx.test.espresso.EspressoException;
+import androidx.test.espresso.UiController;
 import androidx.test.espresso.ViewInteraction;
+import androidx.test.espresso.action.GeneralClickAction;
+import androidx.test.espresso.action.Tapper;
 import androidx.test.espresso.contrib.DrawerActions;
 import androidx.test.espresso.contrib.PickerActions;
 import io.appium.espressoserver.lib.handlers.exceptions.AppiumException;
